cleaing up some warnings; removing experimental work: hbart, zbart

+++ BART3/src/cabart.h
@@ -73,7 +73,7 @@ RcppExport SEXP cabart(
    Rcpp::IntegerVector _nc(_inc);
    int *numcut = &_nc[0];
    size_t nd = Rcpp::as<int>(_ind);
-   size_t burn = Rcpp::as<int>(_iburn);
+   int burn = Rcpp::as<int>(_iburn);
    size_t thin = Rcpp::as<int>(_ithin);
    double mybeta = Rcpp::as<double>(_ipower);
    double alpha = Rcpp::as<double>(_ibase);
@@ -140,7 +140,7 @@ void cabart(
    size_t m,		//number of trees
    int *numcut,		//number of cut points
    size_t nd,		//number of kept draws (except for thinnning ..)
-   size_t burn,		//number of burn-in draws skipped
+   int burn,		//number of burn-in draws skipped
    size_t thin,		//thinning
    double mybeta,
    double alpha,
@@ -211,7 +211,7 @@ void cabart(
    if(np) printf("xp1,xp[np*p]: %lf, %lf\n",ixp[0],ixp[np*p-1]);
    printf("*****Number of Trees: %zu\n",m);
    printf("*****Number of Cut Points: %d ... %d\n", numcut[0], numcut[p-1]);
-   printf("*****burn,nd,thin: %zu,%zu,%zu\n",burn,nd,thin);
+   printf("*****burn,nd,thin: %d,%zu,%zu\n",burn,nd,thin);
 // printf("Prior:\nbeta,alpha,tau,nu,lambda,offset: %lf,%lf,%lf,%lf,%lf,%lf\n",
 //                    mybeta,alpha,tau,nu,lambda,Offset);
    cout << "*****Prior:beta,alpha,tau,nu,lambda,offset: "
